fix(regex): Add negative lookbehind for string
Add negative lookbehind to single line comment regex to make sure the
comment symbol is not part of a string ('#' is a string, not a comment).

Add new function `hashNoCurl()` for Ruby's #{} string syntax.

Also, updated `readMultiLineSame()` to check if the comment starts and
ends in the same line and do not pick a line with comment symbol in
string.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>